Rename `claims` to `entitlements-store`
