+++ containers/entitlements-store/README.md
@@ -1,15 +1,8 @@
-# Claims
+# Entitlements store
 
-The claims service provides `claims` JSON blobs to an identity provider. We
-currently are using Keycloak as the IdP.
+The `entitlements-store` service is a *read-only* service that fetches and returns any entity<->attribute mappings present in the data store (currently Postgres). It is currently consumed as an entitlements input source by the [`entitlement-pdp`](../entitlement-pdp) in authZ flows.
 
-1. Upon successful client authentication via tge IdP, Keycloak, that service
-   runs a plugin to invoke this service.
-2. An instance of the Claims service processes the request. It creates a
-   Claims Object appropriate for that client in the current context,
-   and returns it to the IdP/Keycloak.
-3. Our IdP then returns a signed JWT with the Claims Object
-   inside.
+It stands in contrast to the [`entitlements` service](../entitlements), which is a *read-write* service.
 
 ## Development
 
@@ -52,12 +45,12 @@ docker run --rm -p 5432:5432 -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=myPo
 ### build image
 ```shell
 # from project root
-docker build --no-cache --tag opentdf/claims:0.2.0 claims
+docker build --no-cache --tag opentdf/entitlements-store:0.2.0 claims
 ```
 
 ### secrets
 ```shell
-kubectl create secret generic claims-secrets \
+kubectl create secret generic entitlements-store-secrets \
   --from-literal=POSTGRES_PASSWORD=myPostgresPassword \
   --from-file=KAS_CERTIFICATE=certs/kas-public.pem \
   --from-file=KAS_EC_SECP256R1_CERTIFICATE=certs/kas-ec-secp256r1-public.pem
@@ -66,5 +59,5 @@ kubectl create secret generic claims-secrets \
 ### helm
 ```shell
 # from project root
-helm upgrade --install claims ./charts/claims --debug
+helm upgrade --install entitlements-store ./charts/entitlements-store --debug
 ```
