+++ containers/entitlements-store/VERSION
