+++ containers/entitlements-store/requirements.txt
